Separates types and conversion functions in soltest.
